Восстановлены изменения (558-595), которые были утеряны после падения и последующего поднятия sourceforge.net.
git-svn-id: svn://svn.code.sf.net/p/mpcbe/code/trunk@558 7426f7ed-705a-47fc-8cda-d59144bb0dce

+++ src/DSUtil/DSUtil.cpp
@@ -3303,7 +3303,7 @@ int av_reduce(int *dst_num, int *dst_den,
         den = abs(den) / gcd;
     }
     if (num <= max && den <= max) {
-        a1 = { num, den };
+        a1 = { (int)num, (int)den };
         den = 0;
     }
 
@@ -3318,12 +3318,12 @@ int av_reduce(int *dst_num, int *dst_den,
             if (a1.den) x = min(x, (max - a0.den) / a1.den);
 
             if (den * (2 * x * a1.den + a0.den) > num * a1.den)
-                a1 = { x * a1.num + a0.num, x * a1.den + a0.den };
+                a1 = { int(x * a1.num + a0.num), int(x * a1.den + a0.den) };
             break;
         }
 
         a0  = a1;
-        a1  = { a2n, a2d };
+        a1  = { (int)a2n, (int)a2d };
         num = den;
         den = next_den;
     }
